requirements.txt
Authomatic==0.1.0.post1
Fabric3==1.12.0.post1
Flask==0.11
Flask-Admin==1.4.2
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11
bcdoc==0.12.0
boto==2.34.0
botocore==0.31.0
clan==0.1.3
-e git://github.com/nprapps/copytext.git@411e3d6297bdbc4ca143c964d68deeb757f99ac7#egg=copytext
csvkit==0.9.1
docutils==0.11
elex==2.4.1
gunicorn==19.1.1
httplib2==0.9
jmespath==0.2.1
joblib==0.10.3
nose==1.2.1
odict==1.5.1
openpyxl==2.2.0
peewee==2.8.5
ply==3.4
psycopg2==2.6.2
pyasn1==0.1.7
pycrypto==2.6
python-dateutil==2.2
python-slugify==1.2.1
pytz==2016.7
requests==2.11.1
six==1.10.0
simplejson==3.8.2
slimit==0.7.4
smartypants==1.8.6
ssh==1.7.14
termcolor==1.1.0
wheel==0.23.0
wtf-peewee==0.2.6